refactor: resDto 레이어 이동으로 인한 import문 수정
